refactored operator jobs history as a subroute to operator jobs
